Fix run_tests.sh exiting with the wrong status
and complaining when there are no gcda-files.

+++ run_tests.sh
@@ -1,9 +1,11 @@
 #!/bin/sh
-rm *.gcda
-gcc --coverage -fprofile-exclude-files=test\.c -o test test.c -lm
-status=$?
+[ -e *.gcda ] && rm *.gcda
+gcc --coverage -fprofile-exclude-files=test\.c -o test test.c -lm \
+	|| exit 1
 
 ./test
+status=$?
+
 echo ""
 gcov test
